requirements.txt
kh-common[aerospike,auth,logging,sql]~=0.6.6
aiofiles~=0.5.0
psycopg2-binary~=2.8.5
ujson~=3.0.0
aiohttp~=3.6.2
uvicorn~=0.11.6
gunicorn~=20.0.4
argon2-cffi~=20.1.0
cryptography~=3.0
avrofastapi~=0.0.1